# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Motolegacy",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://KidziBack.squaretech.tech/motolegacy/api"
        },
        {
            "url": "http://127.0.0.1:8000/motolegacy/api"
        }
    ],
    "paths": {
...
        "/cars": {
            "get": {
                "tags": [
                    "User - Cars"
                ],
                "summary": "List all visible cars with optional filters",
                "operationId": "d8232bc6a16fec784f4a5fca541f9856",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sale",
                                "rent"
                            ]
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "User - Cars"
                ],
                "summary": "Get car details",
                "operationId": "8b43987d328663989329b7b6b01db674",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/favorites": {
            "get": {
                "tags": [
                    "User - Favorites"
                ],
                "summary": "List user's favorite cars",
                "operationId": "de458f646b8ab1d9eb829bbd6a218808",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User - Favorites"
                ],
                "summary": "Toggle a car in user's favorites",
                "operationId": "0caf4d36b19c2609b4304ed4df1afee9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "car_id"
                                ],
                                "properties": {
                                    "car_id": {
                                        "description": "ID of the car",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Toggled successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/ratings/{seller_id}": {
            "get": {
                "tags": [
                    "User - Ratings & Reviews"
                ],
                "summary": "List all ratings for a specific seller",
                "operationId": "7bb5eea16adbc6dc7754b0416b559ed0",
                "parameters": [
                    {
                        "name": "seller_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/ratings": {
            "post": {
                "tags": [
                    "User - Ratings & Reviews"
                ],
                "summary": "Rate a seller",
                "operationId": "f76110814bddc4625c9a09750809714c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "seller_id",
                                    "rating"
                                ],
                                "properties": {
                                    "seller_id": {
                                        "description": "ID of the seller",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Rating value from 1 to 5",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Great seller, highly recommended"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating submitted"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}